+++ src/bluesearch/entrypoint/database/add.py
@@ -30,7 +30,7 @@ def get_parser() -> argparse.ArgumentParser:
         help="""
         The location of the database depending on the database type.
 
-        For MySQL the server URL should be provided, for SQLite the
+        For MySQL and MariaDB the server URL should be provided, for SQLite the
         location of the database file. Generally, the scheme part of
         the URL should be omitted, e.g. for MySQL the URL should be
         of the form 'my_sql_server.ch:1234/my_database' and for SQLite
@@ -46,7 +46,7 @@ def get_parser() -> argparse.ArgumentParser:
         "--db-type",
         default="sqlite",
         type=str,
-        choices=("mysql", "sqlite"),
+        choices=("mariadb", "mysql", "sqlite"),
         help="Type of the database.",
     )
     return parser
@@ -73,7 +73,7 @@ def run(
     if db_type == "sqlite":
         engine = sqlalchemy.create_engine(f"sqlite:///{db_url}")
 
-    elif db_type == "mysql":
+    elif db_type in {"mariadb", "mysql"}:
         engine = sqlalchemy.create_engine(f"mysql+pymysql://{db_url}")
 
     else: